renamed config folder to passport, installed morgan and used logger, installed cors, added proxy to client package.json, added logout route, set session in register
