+++ server.js
@@ -4,11 +4,15 @@ const mongoose = require('mongoose');
 const session = require('express-session');
 const cookieParser = require('cookie-parser');
 const passport = require('passport');
+const cors = require('cors');
+const morgan = require('morgan');
 const routes = require('./routes');
 const PORT = process.env.PORT || 3001;
 const app = express();
 
 
+app.use(cors());
+app.use(morgan('dev'));
 
 app.use(express.urlencoded({ extended: true }));
 app.use(express.json());
@@ -23,8 +27,10 @@ app.use(session({
 }));
 
 
+
 app.use(cookieParser('supersecret'));
 
+// require('./passport/passportConfig')(passport);
 app.use(passport.initialize());
 app.use(passport.session());
